py3: test all supported python versions
- Test Python 3.6 again. It was disabled due to
https://github.com/fireeye/capa/issues/405
Let's give it another try. In case there are still problems we would
have to either fix it or not support Python 3.6.
- Test Python 3.9 instead of 3.9.1. According to setup-python
documentation, using patch version will take longer if the exact version
is not already installed in the runner due to more recent version.
Specifying only a major and minor version uses the most recent patch
version, which is the one that we IMO should use.
- Test latest Python version (3.10). This includes both pre-release and
stable versions.