py3: Python 3 knows about cp65001
Python 2 doesn't know about `cp65001`. But Python 3 does.

Note that this code is not equivalent to the previous one in some
versions of Python 3. Since Python 3.8 `cp65001` is an alias to `utf_8`.
But not before Python 3.8 and it used to cause some problems:
https://bugs.python.org/issue36778
Not sure if we should keep this code to ensure the same behavior in
Python < 3.8.