changelog: prepare to automatize new rules entries
Use an empty item in the `New Rules` section as a marker for the GitHub
Action. If this causes problems, we could look into other solution such
as writing 2 lines before `### Bug Fixes`. But I think this is the
easiest I can come up with. So lets give it a try.